Folder Restructuring
the root directory now contains Client and Server in seperate folders
