--- client/Readme.md
@@ -1,88 +1,0 @@
-# Qissa - A New Way For Equality
-
-**Qissa: A Social Media Platform for LGBTQ+ Individuals Worldwide
-Why Qissa?**
-
-**LGBTQ+** individuals face many challenges in their daily lives, such as discrimination, lack of legal awareness, and social isolation. Existing social media platforms may not fully address their unique needs and interests, limiting their opportunities to connect, share, and advocate for change.
-
-Qissa is a social media platform that aims to empower **LGBTQ+** individuals worldwide by providing them with a dedicated space to celebrate their identities, express their opinions, and access relevant information. Qissa is more than just a social media platform; it is a community of support, solidarity, and empowerment.
-What Qissa Offers?
-
-A variety of features and channels to cater to the diverse needs and preferences of LGBTQ+ individuals are offered by **Qissa**.
-
-**_Some of the key features and channels are_**
-
-- **Beautiful Experiences Channel:** A channel where users can share their positive moments and daily joys with other LGBTQ+ individuals. This channel aims to foster a sense of happiness and gratitude among users and celebrate their achievements and milestones.
-
-- **Discrimination Diaries Channel:** A channel where users can discuss and raise awareness about the instances of discrimination they face or witness in their lives. This channel aims to provide a safe space for users to vent their frustrations, seek support, and learn from each other’s experiences.
-
-- **Coming Out Stories Channel:** A channel where users can share their personal coming out stories and offer advice and encouragement to others who are planning to come out. This channel aims to provide a supportive space for users to embrace their identities and overcome their fears and challenges.
-
-- **Legal Advocacy Hub Channel:** A channel where users can facilitate discussions on new LGBTQ+ laws and encourage each other to propose legal changes in their countries. This channel aims to provide a platform for users to advocate for their rights and influence policy-making.
-
-- **General Channel:** A channel where users can engage in general discussions on various topics of interest, such as culture, entertainment, sports, etc. This channel aims to foster community engagement and interaction among users and promote a sense of belonging and friendship.
-
-- **International Legal Resources Feature:** A feature that offers information on LGBTQ+ laws specific to each country, such as marriage, adoption, anti-discrimination, etc. This feature aims to help users navigate legal frameworks confidently and access their rights and protections.
-
-- **Community Mapping Feature:** A feature that connects users with local LGBTQ+ communities, events, and support groups in their areas. This feature aims to help users find and join LGBTQ+ networks and activities near them and expand their social circles.
-
-## How Qissa Works?
-
-**_Qissa is designed with the following principles in mind:_**
-
-- **User-Centric Interface**: The platform prioritizes a user-friendly and customizable interface that allows users to easily navigate and personalize their profiles and settings.
-
-- **Privacy and Security**: The platform implements robust privacy measures and options for anonymous sharing that ensure users’ data and identity are protected and respected.
-
-- **Moderation and Reporting**: The platform ensures a safe and respectful environment with stringent moderation and user reporting tools that prevent and address any abusive or harmful behavior or content.
-
-- **Global Accessibility**: The platform supports multiple languages and provides information in diverse languages that ensure users from different countries and backgrounds can access and enjoy the platform.
-
-## Who Qissa Partners With?
-
-**Qissa** actively seeks collaborations with **LGBTQ+** organizations to enhance the platform’s credibility and impact. Some of the potential partners are:
-
-- **Human Rights Campaign:** A leading LGBTQ+ advocacy group that works to advance the rights and well-being of LGBTQ+ people around the world.
-
-- ILGA World: A global federation of LGBTQ+ organizations that represents more than 1,600 member organizations from 160 countries and regions.
-
-- **OutRight Action International:** An international human rights organization that works to end violence and discrimination against LGBTQ+ people worldwide.
-
-- **GLAAD:** A media advocacy organization that works to promote and ensure fair, accurate, and inclusive representation of LGBTQ+ people and issues in the media.
-
-## How Qissa Improves?
-
-Qissa thrives on user feedback, utilizing surveys, forums, and feedback forms to understand evolving needs. The platform actively incorporates regular updates and enhancements based on user insights, ensuring its dynamic relevance.
-
-Seamless feedback mechanisms and analytics-driven decision-making enable **Qissa** to refine user experience, enhance security, and introduce features aligned with **LGBTQ+ community** preferences.
-
-Transparent communication keeps users informed, fostering a collaborative journey of continuous improvement and innovation.
-
-## WireFrames for the Project Idea
-
-![Screenshot from 2024-01-21 20-56-32](https://github.com/404-GS-IIIT-Kota/Temp/assets/124715224/1df0e2be-baed-4b48-9aae-a9725112bc2c)
-
-## UI For the SignIn Page
-
-![Login](https://github.com/404-GS-IIIT-Kota/Temp/assets/124715224/275086f8-ad85-45c2-9032-c83f9af602de)
-
-## Tech Stack used
-
-- React
-- OAuth
-- Firebase
-- Material UI
-- Nodejs
-- Expressjs,
-- MongoDB
-- RESTApi
-- GoogleAPI
-
-[Qissa PDF File](./Qissa.pdf)
-
-## Team 404
-
-    1. Aaryan Anil Kumar Singh
-    2. Dhairya Amrish Majmudar
-    3. Hiteshwar Kaushik
-    4. Rohan
