+++ client/package.json
@@ -15,6 +15,8 @@
     "@material-tailwind/react": "^2.1.8",
     "@mui/icons-material": "^5.15.6",
     "@mui/material": "^5.15.6",
+    "axios": "^1.6.7",
+    "emoji-picker-react": "^4.7.15",
     "firebase": "^10.8.0",
     "react": "^18.2.0",
     "react-datepicker": "^6.1.0",
