+++ client/src/pages/BeautifulExperiences.jsx
