+++ client/src/pages/LegalAdvocacyHub.jsx
