+++ client/src/pages/Signin.jsx
@@ -5,15 +5,17 @@ import FacebookIcon from "@mui/icons-material/Facebook";
 import TwitterIcon from "@mui/icons-material/Twitter";
 import placeholderImage from "../assets/sign-up-image-1.jpg";
 import { Link } from "react-router-dom";
+import { useState, useEffect } from "react";
 
 const Signup = () => {
+  const [post, setPost] = useState({});
+
   return (
     <div className="h-screen flex items-center justify-center bg-gray-200">
       <div
         className="w-80 h-80 bg-white mx-auto flex flex-col lg:flex-row rounded-2xl shadow-2xl"
         style={{ width: "85%", height: "85%" }}
       >
-        {/* Left Section - Image */}
         <div className="w-full lg:w-1/2 p-6 flex flex-col items-center justify-center">
           <img
             src={placeholderImage}
@@ -26,7 +28,6 @@ const Signup = () => {
           </Link>
         </div>
 
-        {/* Right Section - Signup Form */}
         <div className="w-full lg:w-1/2 p-6 bg-white rounded-lg flex flex-col justify-center">
           <h2 className="text-5xl font-bold mb-4 max-sm:text-center">
             Sign In
@@ -56,14 +57,7 @@ const Signup = () => {
               />
             </div>
 
-            <div className="mb-6 flex items-center max-sm:justify-center">
-              <input type="checkbox" className="mr-2" />
-              <span className="text-gray-700 max-sm:text-center">
-                Remember Me
-              </span>
-            </div>
-
-            <Link to="/main">
+            <Link to="/beautiful-experiences">
               <button
                 type="submit"
                 className="w-2/5  bg-blue-500 text-white py-2 rounded-md hover:bg-blue-600 flex max-sm:items-center max-sm:w-[60%] justify-center max-sm:translate-x-[4rem]"
