+++ client/tailwind.config.js
@@ -12,5 +12,19 @@ export default {
       },
     },
   },
-  plugins: [],
+  plugins: [
+    function ({ addUtilities }) {
+      const newUtilities = {
+        ".no-scrollbar::-webkit-scrollbar": {
+          display : "none",
+        },
+        ".no-scrollbar":{
+          "-ms-overflow-style":"none",
+          "scrollbar-width":"none",
+        },
+      };
+
+      addUtilities(newUtilities);
+    }
+  ],
 }
