+++ client/vite.config.js
@@ -3,5 +3,10 @@ import react from '@vitejs/plugin-react'
 
 // https://vitejs.dev/config/
 export default defineConfig({
+  server: {
+    proxy: {
+      '/api': 'http://localhost:8000'
+    }
+  },
   plugins: [react()],
 })
