--- src/component/Datepicker.jsx
@@ -1,20 +1,0 @@
-import { useState } from "react";
-import DatePicker from "react-datepicker";
-import "react-datepicker/dist/react-datepicker.css";
-
-const Datepicker = () => {
-  const [selectedDate, setSelectedDate] = useState(null);
-
-  return (
-    <div className="relative w-2/5 max-sm:w-full">
-      <DatePicker
-        selected={selectedDate}
-        onChange={(date) => setSelectedDate(date)}
-        className="w-full px-4 py-2 ml-5 max-sm:ml-0 border rounded-md focus:outline-none focus:border-blue-500 border-gray-300"
-        placeholderText="Your Birthday"
-      />
-    </div>
-  );
-};
-
-export default Datepicker;
