--- src/component/Loader.jsx
@@ -1,5 +1,0 @@
-const Loader = () => {
-  return <div>Loader</div>;
-};
-
-export default Loader;
