--- src/pages/Temp.jsx
@@ -1,6 +1,0 @@
-
-const Temp = () => {
-  return <div>Temp</div>;
-};
-
-export default Temp;
