--- tailwind.config.js
@@ -1,30 +1,0 @@
-/** @type {import('tailwindcss').Config} */
-export default {
-  content: [
-    "./index.html",
-    "./src/**/*.{js,ts,jsx,tsx}",
-  ],
-  theme: {
-    extend: {
-      height: {
-        100 : "83vh",
-        900 : "90%",
-      },
-    },
-  },
-  plugins: [
-    function ({ addUtilities }) {
-      const newUtilities = {
-        ".no-scrollbar::-webkit-scrollbar": {
-          display : "none",
-        },
-        ".no-scrollbar":{
-          "-ms-overflow-style":"none",
-          "scrollbar-width":"none",
-        },
-      };
-
-      addUtilities(newUtilities);
-    }
-  ],
-}
