--- vite.config.js
@@ -1,12 +1,0 @@
-import { defineConfig } from 'vite'
-import react from '@vitejs/plugin-react'
-
-// https://vitejs.dev/config/
-export default defineConfig({
-  server: {
-    proxy: {
-      '/api': 'http://localhost:8000'
-    }
-  },
-  plugins: [react()],
-})
